updated drivers_test_unit.py to automatically identify and test for settings attributes, relocated TestInstrumentDriver to it's own file under drivers, and updated TestInstrumentDriver to delete bad settings attribute after testing for failure upon attempted creation.
